requirements.txt
pandas==2.0.3
geopandas==0.13.2
folium==0.14.0
plotly==5.15.0
dash==2.11.1
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
requests==2.31.0
beautifulsoup4==4.12.2
fiona==1.9.4
shapely==2.0.1
rasterio==1.3.8
networkx==3.1
scikit-learn==1.3.0
flask==2.3.2
flask-cors==4.0.0
jupyterlab==4.0.3
matplotlib==3.7.2
seaborn==0.12.2
openpyxl==3.1.2
joblib==1.3.1